+++ pkg/capybara/setup_test.go
@@ -1,0 +1,17 @@
+package capybara
+
+import "testing"
+
+// This file is only here to setup tools for testing purpose
+
+func assertEqual(t *testing.T, a interface{}, b interface{}) {
+	if a != b {
+		t.Fatalf("%s != %s", a, b)
+	}
+}
+
+func assertNotEqual(t *testing.T, a interface{}, b interface{}) {
+	if a == b {
+		t.Fatalf("%s != %s", a, b)
+	}
+}