+++ CHANGELOG.md
@@ -1,1 +1,11 @@
+### 1.1 - 2019/11/04
+- Add "method" key inside "services", in config JSON, allowing the use of different methods of pattern matching.
+- Add "string" and "regex" method of pattern matching. "string" is more simple and faster
+- Add tests for the method system
+- Add tests for handler and config
+- Move $root/*.go files into $root/pkg/capybara
+- Update README.md
+- Update go.mod
+- Update go version in travis.yml
+
 ### 1.0